fix(build): dynamic build date + preview config + audit constants
Vite previously embedded a hardcoded build date ("20260216") and shipped
demo user fixtures with literal "password" entries from the application's
constants module. The preview server also had no explicit port, so `vite
preview` defaulted to 4173 instead of matching the dev port.

- vite.config.ts: derive YYYYMMDD from `new Date()` at config-eval time
  (extracted to scripts/buildDate.ts for testability) and add a preview
  server stanza pinned to port 3000 / host 0.0.0.0 to match dev.
- constants.tsx: strip the unused DEFAULT_USERS/CLIENTS/PROJECTS/TASKS
  fixtures (only COLORS was imported anywhere). The demo data, including
  the placeholder "password" strings, moves to test/fixtures/defaultData.ts
  with a header comment marking it as test/demo-only.
- Tests: unit-test formatBuildDate / getBuildDate (covers zero-padding
  and the regression against the old static literal) and assert that
  constants.tsx no longer exports the user-fixture symbols.
